Introduce the lh5.core module with the main functions (read,write,...)
