+++ src/lgdo/lh5/__init__.py
@@ -6,6 +6,7 @@ browsed easily in python like any `HDF5 <https://www.hdfgroup.org>`_ file using
 """
 from __future__ import annotations
 
+from .core import read
 from .iterator import LH5Iterator
 from .store import LH5Store
 from .tools import load_dfs, load_nda, ls, read_as, show
@@ -16,6 +17,7 @@ __all__ = [
     "LH5Store",
     "load_dfs",
     "load_nda",
+    "read",
     "read_as",
     "ls",
     "read_n_rows",
